requirements.txt
fastapi==0.128.0
uvicorn==0.38.0
httpx==0.28.1
PyYAML==6.0.3
pydantic==2.11.7